Move the interprocess communication to ipc.py
